feat: add global exception handling and OpenAPI integration
- Implemented a global exception handler to centralize error management
- Added custom error messages and detailed responses for each controller
- Displayed request path in error responses only in dev environment (hidden in prod)
- Moved exception handling logic from controllers to service layer for cleaner code structure
- Converted Postman collection to enhanced openapi.yaml for Swagger API documentation

+++ src/main/java/com/codingJava/LibraryApp/service/BookService.java
@@ -1,6 +1,7 @@
 package com.codingJava.LibraryApp.service;
 
 import com.codingJava.LibraryApp.entity.Book;
+import com.codingJava.LibraryApp.exception.ResourceNotFoundException;
 import com.codingJava.LibraryApp.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,18 +19,33 @@ public class BookService {
     }
 
     public Book getBookByName(String bookName) {
-       Book bookFound =  bookRepository.findBookByTitle(bookName);
-       return bookFound;
+        Book bookFound = bookRepository.findBookByTitle(bookName);
+        if (bookFound == null) {
+            throw new ResourceNotFoundException(
+                    "Book not found with name: " + bookName,
+                    "The book you're searching for may have been deleted or never existed."
+            );
+        }
+        return bookFound;
+    }
+    public Book getBookById(Integer id) {
+        return bookRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(
+                "Book not found with Id: " + id,
+                "The book you're searching for may have been deleted or never existed."
+        ));
     }
 
     public List<Book> getAllBooks() {
-       List<Book> books =  bookRepository.findAll();
-       return books;
+        return bookRepository.findAll();
     }
 
 
     public Book updateBook(Integer id, Book book){
-        Book bookFound = bookRepository.findById(id).orElseThrow(()-> new RuntimeException("Book not found"));
+        Book bookFound = bookRepository.findById(id)
+                .orElseThrow(()->
+                        new ResourceNotFoundException(
+                                "Book with id "+ id +" not found",
+                                "The book you're searching for may have been deleted or never existed."));
         bookFound.setTitle(book.getTitle());
         bookFound.setAuthor(book.getAuthor());
         bookFound.setGenre(book.getGenre());
@@ -37,6 +53,12 @@ public class BookService {
     }
 
     public void deleteBook(Integer id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Book not found with ID: " + id,
+                    "Cannot delete. The book with this ID does not exist."
+            );
+        }
         bookRepository.deleteById(id);
     }
